Remove DataPipeline
