+++ yolort/data/__init__.py
@@ -1,4 +1,3 @@
 # Copyright (c) 2021, Zhiqiang Wang. All Rights Reserved.
-from .coco_eval import COCOEvaluator
-from .data_pipeline import DataPipeline
 from .data_module import DetectionDataModule, VOCDetectionDataModule, COCODetectionDataModule
+from .coco_eval import COCOEvaluator
